--- src/strategies/erc20-balance-of-minto-all/examples.json
@@ -1,35 +1,0 @@
-[
-  {
-    "name": "Example query",
-    "strategy": {
-      "name": "erc20-balance-of",
-      "params": {
-        "address": "0x6b175474e89094c44da98b954eedeac495271d0f",
-        "symbol": "DAI",
-        "decimals": 18
-      }
-    },
-    "network": "1",
-    "addresses": [
-      "0xa478c2975ab1ea89e8196811f51a7b7ade33eb11",
-      "0xeF8305E140ac520225DAf050e2f71d5fBcC543e7",
-      "0x1E1A51E25f2816335cA436D65e9Af7694BE232ad",
-      "0x1F717Ce8ff07597ee7c408b5623dF40AaAf1787C",
-      "0x1c7a9275F2BD5a260A9c31069F77d53473b8ae2e",
-      "0x1d5E65a087eBc3d03a294412E46CE5D6882969f4",
-      "0x1f254336E5c46639A851b9CfC165697150a6c327",
-      "0x2ec3F80BeDA63Ede96BA20375032CDD3aAfb3030",
-      "0x4AcBcA6BE2f8D2540bBF4CA77E45dA0A4a095Fa2",
-      "0x4F3D348a6D09837Ae7961B1E0cEe2cc118cec777",
-      "0x6D7f23A509E212Ba7773EC1b2505d1A134f54fbe",
-      "0x07a1f6fc89223c5ebD4e4ddaE89Ac97629856A0f",
-      "0x8d5F05270da470e015b67Ab5042BDbE2D2FEFB48",
-      "0x8d07D225a769b7Af3A923481E1FdF49180e6A265",
-      "0x8f60501dE5b9b01F9EAf1214dbE1924aA97F7fd0",
-      "0x9B8e8dD9151260c21CB6D7cc59067cd8DF306D58",
-      "0x17ea92D6FfbAA1c7F6B117c1E9D0c88ABdc8b84C",
-      "0x38C0039247A31F3939baE65e953612125cB88268"
-    ],
-    "snapshot": 11437846
-  }
-]
