--- src/strategies/erc20-balance-of-minto-all/schema.json
@@ -1,33 +1,0 @@
-{
-  "$schema": "http://json-schema.org/draft-07/schema#",
-  "$ref": "#/definitions/Strategy",
-  "definitions": {
-    "Strategy": {
-      "title": "Strategy",
-      "type": "object",
-      "properties": {
-        "symbol": {
-          "type": "string",
-          "title": "Symbol",
-          "examples": ["e.g. UNI"],
-          "maxLength": 16
-        },
-        "address": {
-          "type": "string",
-          "title": "Contract address",
-          "examples": ["e.g. 0x1f9840a85d5aF5bf1D1762F925BDADdC4201F984"],
-          "pattern": "^0x[a-fA-F0-9]{40}$",
-          "minLength": 42,
-          "maxLength": 42
-        },
-        "decimals": {
-          "type": "number",
-          "title": "Decimals",
-          "examples": ["e.g. 18"]
-        }
-      },
-      "required": ["address", "decimals"],
-      "additionalProperties": false
-    }
-  }
-}
