+++ src/strategies/index.ts
@@ -25,7 +25,7 @@ import * as erc20WithBalance from './erc20-with-balance';
 import * as erc20BalanceOfDelegation from './erc20-balance-of-delegation';
 import * as erc20BalanceOfQuadraticDelegation from './erc20-balance-of-quadratic-delegation';
 import * as erc20BalanceOfWeighted from './erc20-balance-of-weighted';
-import * as erc20BalanceOfStaked from './erc20-balance-of-staked';
+import * as mintoBalanceAll from './minto-balance-of-all';
 import * as erc20BalanceOfIndexed from './erc20-balance-of-indexed';
 import * as revest from './revest';
 import * as erc20Price from './erc20-price';
@@ -325,7 +325,7 @@ const strategies = {
   'erc20-balance-of-delegation': erc20BalanceOfDelegation,
   'erc20-balance-of-quadratic-delegation': erc20BalanceOfQuadraticDelegation,
   'erc20-balance-of-weighted': erc20BalanceOfWeighted,
-  'erc20-balance-of-staked': erc20BalanceOfStaked,
+  'minto-balance-of-all': mintoBalanceAll,
   'erc20-balance-of-indexed': erc20BalanceOfIndexed,
   'erc20-price': erc20Price,
   'balance-of-with-min': balanceOfWithMin,
