+++ src/strategies/minto-balance-of-all/examples.json
@@ -2,7 +2,7 @@
   {
     "name": "Example query",
     "strategy": {
-      "name": "erc20-balance-of-staked",
+      "name": "minto-balance-of-all",
       "params": {
         "address": "0x410a56541bD912F9B60943fcB344f1E3D6F09567",
         "symbol": "BTCMT",
@@ -16,7 +16,16 @@
       "0x05903a958710ec752dd4a25dd3157ccee9d6c92e",
       "0x03bb74c32acb025effab4d31e579d3751363c216",
       "0x0369a66d5d25734a9aef4689546a809888fe5bfe",
-      "0x0743a7eb79cf4b478ba0bba22bdf2f11bc8ad5de"
+      "0x0743a7eb79cf4b478ba0bba22bdf2f11bc8ad5de",
+     "0x00ce049d0624a087473a4bd5078ac90ad25175e8",
+      "0x01f3a3962668df7799429ae827df59bc32a5d761",
+      "0x020ce48bb83cd6cb1483149ed7a6d2bca4526c84",
+      "0x0235ac7285d4ad1a5261904539e4862c13a830af",
+      "0x0369a66d5d25734a9aef4689546a809888fe5bfe",
+      "0x03bb74c32acb025effab4d31e579d3751363c216",
+      "0x05903a958710ec752dd4a25dd3157ccee9d6c92e",
+      "0x0636b47164ea3099c6fc39c52a3f4b7b7ecaa03b",
+      "0x0686b62cd088715dcb690e6514aaeca229bbac07"
     ],
     "snapshot": 11437846
   }
